Fold `TagHelperDescriptorMatcher` into `TagHelperBlockRewriter`
